% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_shapefiles.R
\name{all_shapefiles}
\alias{all_shapefiles}
\title{Return path to all shapefiles}
\usage{
all_shapefiles(check_dl = FALSE, dataset = c("nhdh", "hydrolakes",
  "nhdplusv2"), feature_type = c("waterbody", "flowline"))
}
\arguments{
\item{check_dl}{If TRUE, checks to ensure all files for that dataset have been downloaded.
This check takes some time (~30 seconds) to check all files (and much longer to dowload if necessary)}

\item{dataset}{which dataset to return shapefiles from.}

\item{feature_type}{return waterbodies or flowlines. The hydrolakes dataset does not include flowlines.}
}
\description{
Returns list of paths to shapefiles for use in custom processing.
}
