% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_path_helpers.R
\name{cache_get_dir}
\alias{cache_get_dir}
\title{Get local cache path}
\usage{
cache_get_dir()
}
\description{
Downloaded shapefiles and databases are cached locally.
This location can be accessed and reset so that the best
available locaiton can be used. For example, a fast SSD
or large secondary hard drive may be used.
}
\examples{
#see where cached files are being stored
print(cache_get_dir())


}
\seealso{
\code{\link{cache_set_dir}}
}
