% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_info.R
\name{dataset_info}
\alias{dataset_info}
\title{Return all info for requested dataset}
\usage{
dataset_info(dataset, feature_type)
}
\arguments{
\item{dataset}{Name of dataset}

\item{feature_type}{Feature type (flowline or waterbody)}
}
\value{
List with dataset metadata elements including
\describe{
  \item{bb_cache_path}{Path to Rdata file with bounding box cache for dataset}
  \item{shapefile_name}{Dataset-unique name for region shapefiles}
  \item{db_path}{Path to sqlite database for shape-by-ID lookup}
  \item{file_index_path}{Path to the dataset's download file index and hash lookup}
  \item{id_column}{Dataset-unique ID column name}
}
}
\description{
Single point
}
