% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_by_id.R
\name{flowline_shape_by_id}
\alias{flowline_shape_by_id}
\title{Link IDs to flowline shapefiles}
\usage{
flowline_shape_by_id(PERMANENT_match)
}
\arguments{
\item{PERMANENT_match}{filter using PERMANENT_}
}
\value{
SpatialPointsDataFrame containing polygons with associated IDs.
}
\description{
Get shapefiles containing flowlines with specified IDs.
}
