% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_shape_waterbodies.R
\name{link_shape_waterbodies}
\alias{link_shape_waterbodies}
\title{Link geoshapes to Waterbodies}
\usage{
link_shape_waterbodies(shapedf)
}
\arguments{
\item{shapedf}{SpatialPolygonsDataFrame of shapes to match}
}
\value{
matching waterbody attributes
}
\description{
Link geoshapes to waterbodies
}
