% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_streams.R
\name{link_streams}
\alias{link_streams}
\title{Link geopoints to streams}
\usage{
link_streams(lats, lons, ids, max_dist = 100)
}
\arguments{
\item{lats}{Vector of point latitudes}

\item{lons}{Vector of point longitudes}

\item{ids}{Vector of point identifiers (string or numeric)}
}
\value{
flowline permanent ids
}
