% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_to_nhd.R
\name{link_to_waterbodies}
\alias{link_to_waterbodies}
\title{Link geopoints to Waterbodies}
\usage{
link_to_waterbodies(lats, lons, ids, dataset = "nhd", buffer = 0)
}
\arguments{
\item{lats}{Vector of point latitudes}

\item{lons}{Vector of point longitudes}

\item{ids}{Vector of point identifiers (string or numeric)}

\item{dataset}{Character name of dataset to link against. Can be either "nhd" or "hydrolakes"}
}
\value{
Water body permanent IDs
}
\description{
Link geopoints to waterbodies in a geospatial dataset.
}
