% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_path_helpers.R
\name{local_path}
\alias{local_path}
\title{Get local cache path}
\usage{
local_path()
}
\description{
Downloaded shapefiles and databases are cached locally.
This location can be accessed and reset so that the best
available locaiton can be used. For example, a fast SSD
or large secondary hard drive may be used.
}
\examples{
#see where cached files are being stored
print(local_path())


}
\seealso{
\code{\link{set_local_path}}
}
