% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{local_storage}
\alias{local_storage}
\title{Path to local NHD storage}
\usage{
local_storage()
}
\description{
Returns internal cache path to local NHD storage
location. Can be set by \link{set_local_storage}.
}

