% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dl_file.R
\name{set_local_files_path}
\alias{set_local_files_path}
\title{Set local files path}
\usage{
set_local_files_path(path = NULL)
}
\arguments{
\item{path}{character path to new local files path. If null, path will be reset to default user data directory location.}
}
