% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{set_local_storage}
\alias{set_local_storage}
\title{Set path to local NHD storage}
\usage{
set_local_storage(storage_path)
}
\arguments{
\item{storage_path}{Character value of new
local storage path}
}
\description{
Sets internal cache path to local NHD storage
location. Can be retrieved by \link{local_storage}.
}
