% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_streams.R
\name{snapPointsToLines}
\alias{snapPointsToLines}
\title{Link geoshapes to streams}
\usage{
snapPointsToLines(points, lines, maxDist = NA, withAttrs = TRUE,
  idField = NA)
}
\arguments{
\item{shapedf}{SpatialPolygonsDataFrame of shapes to match}
}
\value{

}
