% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_flowlines.R
\name{traverse_flowlines}
\alias{traverse_flowlines}
\title{traverse_flowlines}
\usage{
traverse_flowlines(distance, start, direction = c("out", "in"),
  max_depth = 10000)
}
\arguments{
\item{distance}{maximum distance to traverse in km. If negative, traverse until the ocean (node 0) or max_depth is reached.}

\item{start}{character node to start}

\item{direction}{character; either "out" or "in"}

\item{max_depth}{maximum traversal depth before terminating}
}
\value{
list of nodes traversed
}
\description{
traverse hydrological network
}
\examples{
\dontrun{
traverse_flowlines(1000, "141329377", "out")
}
}
