% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_by_id.R
\name{waterbody_shape_by_id}
\alias{waterbody_shape_by_id}
\title{Link IDs to warterbody shapefiles}
\usage{
waterbody_shape_by_id(PERMANENT_match = NULL, GNIS_ID_match = NULL,
  GNIS_NAME_match = NULL, REACHCODE_match = NULL)
}
\arguments{
\item{PERMANENT_match}{filter using PERMANENT_}

\item{GNIS_ID_match}{filter using GNIS_ID}

\item{GNIS_NAME_match}{filter using GNIS_NAME}

\item{REACHCODE_match}{filer using REACHCODE}
}
\value{
data table linking ids to filenames which contain them.
}
