% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_path_helpers.R
\name{cache_info}
\alias{cache_info}
\title{Get local file cache info}
\usage{
cache_info()
}
\description{
prints files stored at the current cache location and their size in megabytes.
}
\examples{
\dontshow{cache_set_dir(temppath=TRUE)}
cache_info()

}
