% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_flowlines.R
\name{traverse_flowlines}
\alias{traverse_flowlines}
\title{traverse_flowlines}
\usage{
traverse_flowlines(g, distance, start, direction = c("out", "in"),
  max_depth = 10000)
}
\arguments{
\item{g}{network graph - load using \code{\link[dplyr]{src_sqlite}}}

\item{distance}{maximum distance to traverse in km}

\item{start}{character node to start}

\item{direction}{character; either "out" or "in"}
}
\value{
list of nodes traversed
}
\description{
traverse hydrological network
}
\examples{
\dontrun{
traverse_flowlines(src_sqlite("flowtable.sqlite3"), 1000, "141329377", "out")
}
}
